% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnergyRank.R
\name{print.recp}
\alias{print.recp}
\title{Print method for recp objects}
\usage{
\method{print}{recp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"recp"}.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print method for recp objects
}
